[REFACTOR] 도메인 구조 설계
