+++ src/main/java/racing/domain/car/vo/fuel/RandomFuel.java
@@ -1,4 +1,4 @@
-package racing.domain.fuel;
+package racing.domain.car.vo.fuel;
 
 import util.RandomUtils;
